==> Scripts/3_Feature_Selection_Final.ipynb <==
# Feature Selection
## Load the train_reduced and test_reduced data from google drive

# Code to read csv file into Colaboratory:
!pip install -U -q PyDrive
from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials
# Authenticate and create the PyDrive client.
auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)

## Import package

import pandas as pd
import numpy as np
import sklearn
from sklearn.feature_selection import SelectKBest
import statsmodels.api as sm
import random

## Load train_reduced

The shape of train_reduced should be (68381, 523)

link = 'https://drive.google.com/open?id=18JdiygGsvo9ojuV-qWhn2V40HMMbwcjg'
fluff, id = link.split('=')
print (id) # Verify that you have everything after '='

downloaded = drive.CreateFile({'id':id}) 
downloaded.GetContentFile('train_reduced_va.csv')  
df_train = pd.read_csv('train_reduced_va.csv')
print(df_train.shape)


## Load test_reduced

The shape of train_reduced should be (29307, 523)

link = 'https://drive.google.com/open?id=1FILBSURi1WGVoO-8VNu6i2D0LV3NAmqz'
fluff, id = link.split('=')
print (id) # Verify that you have everything after '='

downloaded = drive.CreateFile({'id':id}) 
downloaded.GetContentFile('test_reduced_va.csv')  
df_test = pd.read_csv('test_reduced_va.csv')
print(df_test.shape)

## Use train dataset to select the features
Rename the training data to df3 for avoiding reload the original dataset from the beginning. 

df3 = df_train

Remove `ee` from the train data and create a new dataframe called df4

df4 = df3.loc[:,df3.columns != 'ee']
df4.head()

## Identify x and y and fit the OLS model
Keep `oldmodel` and other crime and lightning variables in x, `ncat` as y


x = df4.loc[:,df4.columns != 'ncat']
y = df4.loc[:,'ncat']
#Adding constant column of ones, mandatory for sm.OLS model
X_1 = sm.add_constant(x)
#Fitting sm.OLS model
model = sm.OLS(y,X_1).fit()
model.pvalues

## Loop over the p-value from the model and eliminate those insignificant variales(p-value > 0.05)

#Backward Elimination
random.seed(45)
cols = list(x.columns)
pmax = 1
while (len(cols)>0):
    p= []
    X_1 = x[cols]
    X_1 = sm.add_constant(X_1)
    model = sm.OLS(y,X_1).fit()
    p = pd.Series(model.pvalues.values[1:],index = cols)      
    pmax = max(p)
    feature_with_p_max = p.idxmax()
    if(pmax>0.05):
        cols.remove(feature_with_p_max)
    else:
        break
selected_features_BE = cols
print(selected_features_BE)

va_withold= selected_features_BE
len(va_withold)

### Extract from Train

# create indices
indices_withold = list(va_withold)
selected_withold = df_train.loc[:, indices_withold]


len(selected_withold.columns)

Selected total 34 features from the model

### Feature treatment (using winsorize to cut in 99th persentile)

## winsorize result_cr
from scipy import stats
for col in selected_withold.columns:
  selected_withold[col] = stats.mstats.winsorize(selected_withold[col], limits=[0.01,0.01], inplace=True) 

selected_withold.describe()

### Add `ncat` and `ee` to the new dataframe called selected_34, and write to csv file and rename the file name 'selected_34_tr.csv'.

## add ncat and ee and oldmodel in the result_lgt
selected_34 = selected_withold
selected_34['ncat'] = df4['ncat']
selected_34['oldmodel']=df4['oldmodel']
selected_34['ee']=df3['ee']
print(selected_34)

## write the result_cr1 to csv
selected_34.to_csv('selected_34_tr.csv',index = False)

### Extract 34 selected features from test_reduced data and create a new dataframe with 34 features called df7



# create indices
indices_withold = list(va_withold)
selected_withold = df_test.loc[:, indices_withold]

### Feature treatment (using winsorize to cut in 99th persentile)

## winsorize result_cr
from scipy import stats
for col in selected_withold.columns:
  selected_withold[col] = stats.mstats.winsorize(selected_withold[col], limits=[0.01,0.01], inplace=True) 

selected_withold.describe()

### Add `ncat` and `ee` to the new dataframe called selected_34, and write to csv file and rename the file name 'selected_34_te.csv'.

## add ncat and ee and oldmodel in the result_lgt
selected_34 = selected_withold
selected_34['ncat'] = df_test['ncat']
selected_34['oldmodel']=df_test['oldmodel']
selected_34['ee']=df_test['ee']
print(selected_34)

## write the result_cr1 to csv
selected_34.to_csv('selected_34_te.csv',index = False)
